requirements.txt
flask
requests
pandas
vaderSentiment
google-generativeai
python-dotenv
